TabbedContainer : Fix tab title truncation issues.
When we used margins to inset the tab bar, after a tab re-arrangement,
the title text of the first/last tab would often be truncated within the
tab header.

Switching to a position offset fixes this, but produces a slight underlap
on the opposite side, under the pinning/panel menus. Attempts to remedy
this with right-based CSS tricks failed, as such, we revert to simply
'covering it with a rug'.