CompoundEditor : Save/Load detached panels and window positions in Layouts
This is a first-pass implementation that makes the sweeping statement that
all properties that are stored should always be recalled. There are variety
of situations where this may - or may not - be ideal. Always doing it is a
simper initial strategy which will hopefully surface real-world issues
fairly quickly. It'd be good to get some user testing on this to identify
and opportunities to make it more intuitive/useful.

State is serialised using bottom-left NDC coordinates relative to each
screens available area. This should hopefully make for a more portable
definition (bottom-left is used over top to better match other Gaffer
conventions). We don't make use of the native Qt save/restore geometry
mechanism as it produces an opaque binary state.

The main downside to this implementation is that (as far as I can tell),
a QWidget needs to be visible in order to determine which Screen it is
on (as otherwise its windowHandle is empty). As such we can't restore
bounds properly during __init__ so it must be called in an appropriate
place after it has been placed on screen.

A possible improvement to this could be to extend the GafferUI.Window API
to cover the extra functionality required to support save/restore, but for
now, we keep some direct interaction with Qt itself.

User Facing Changes :

 - Layouts now store the size, position and fullscreen state of Gaffer
   windows, including detached panels windows.

 - When a layout is restored, window positions will be adjusted.

Side Effects :

 - Default Gaffer layouts now open Maximized, with sensible sizes set if
   the window is un-maximized. This has the knock-on effect that docs
   screenshots are now taken from a bigger window. We need to assess if
   this is ok, or we should adjust the layout/size in screengrab.py.