Editor : Add restorationID to allow layouts to reference a specific editor
Editors have no stable identifier or persistent means of addressing a
specific instance. In order to save UI state held outside of an editor, we
need some way of finding the same editor again after a repr/eval cycle.

This commit introduces `Editor._restorationID()` that is assigned on
construction, and persisted through repr/eval. As the ID only needs to be
unique within the scope of a single layout, it is OK that editors
with the same restorationID may exist in multiple layouts.