IECoreGLPreview : Add frustum type to visualiser rendering
The `VisualisationType::Frustum` is for in-world representations of
the projection/frustum of an object. This could include a camera
frustum + clipping planes, or the cone of a spot light.

We deliberately introduce this as a visualiser type, rather than
controlling it via attributes as we do for some other visualiser
components, as this allows `ObjectVisualiser` implementations to
make use of it too (they don't have access to attributes).

API
---

 - IECoreGLPreview : Adds `VisualisationType:Frustum` for in-scene
   projection/frustum representations.