SpreadsheetUI : Allow rows to be drag-reordered
Shows the vertical section header, and presents this as a drag handle
to allow individual row reordering.

This is somewhat more fiddly than expected, as reordering sections in
QHeaderView is intended as a display transform, but we want to apply
it to the model. This requires us to re-align logical and visual indices
for header sections after the model has been updated to ensure data is
shown in the correct place.