SpreadsheetUI : Improve mouse/<return> key interaction with checkboxes
Hard to get this one right, the double-click-also-toggling the checkbox
problem felt a bit broken. This hopefully feels a bit more intuitive
without going as far as removing single-click interaction from the
checkbox. That would probably be the most consistent thing to do, but
needs some investigation from the Qt side of things.